// src/generated/models/PublicUserProfileDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { BadgeDto } from './BadgeDto';
export type PublicUserProfileDTO = {
    studentId?: string;
    firstName?: string;
    lastName?: string;
    otherName?: string;
    profileImage?: string;
    schoolName?: string;
    totalPracticedQuestion?: number;
    timeSpent?: number;
    accuracy?: number;
    currentStreak?: number;
    lastSeen?: string;
    globalRank?: number;
    totalPoints?: number;
    currentBadge?: BadgeDto;
    ninVerified?: boolean;
};