// src/generated/models/ReferralDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ReferralDTO = {
    id?: string;
    referrerId?: string;
    referralCode?: string;
    referredEmail?: string;
    referredName?: string;
    referredQuestionsAttempts?: number;
    createdAt?: string;
    usedAt?: string;
    totalClaimed?: number;
    totalUnclaimed?: number;
    totalUsed?: number;
    totalUnused?: number;
    used?: boolean;
    claimed?: boolean;
};